requirements.txt
python-telegram-bot==13.15
psycopg2-binary==2.9.9
flask==2.2.5
requests
pytz
python-dotenv
pymongo
telegram